test cases
1)
Enter a number : 14567
The reverse of the given number is: 76541
2)
Enter a number : -45721
The reverse of the given number is: -12754
3)
Enter a number : 
000
The reverse of the given number is: 0
4)
Enter a number : 
AD1947
Invalid entry.. Please enter a number!
5)
Enter a number : !@#$%
Invalid entry.. Please enter a number! 
6)
Enter a number : 
145*999
Invalid entry.. Please enter a number!