test case
1)
Enter the no. of Days :
756
No. of YEARS : 2
No. of WEEKS : 3
No. of DAYS : 5
2)
Enter the no. of Days :
38
No. of YEARS : 0
No. of WEEKS : 5
No. of DAYS : 3
3)
Enter the no. of Days :
3.6
No. of YEARS : 0
No. of WEEKS : 0
No. of DAYS : 3
4)
Enter the no. of Days :
0
INVALID INPUT ! No. of Days Cannot be Zero Or Negative !
5)
C:\Users\User\Desktop>java days
Enter the no. of Days :
-365
INVALID INPUT ! No. of Days Cannot be Zero Or Negative !
6)
Enter the no. of Days :
-45
INVALID INPUT ! No. of Days Cannot be Zero Or Negative !